// src/service/models/IPatient.tsx
export interface IFormInput {
    firstName: string;
    lastName: string;
    preferredName: string;
    address: string;
    primaryNumber: string;
    secondaryNumber: string;
    evaluation: string;
    frequency: string;
    discharge: string;
    recert: string;
    thirtyDaysRelEval: string;
    notes: string;
    userId: number;
    mdNumber: string;
    mdName: string;
    careTeamId: number;
    teamLeader: any;
    ot: any;
    ota: any;
    pt: any;
    pta: any;
    slp: any;
    sn: any;
    aid: any;
    msw: any;
    teamLeaderName: string;
    otname: string;
    otaName: string;
    ptName: string;
    ptaName: string;
    snName: string;
    slpName: string;
    aidName: string;
    mswName: string;
    noOfWeeks: any;
    visitsPerWeek: any;
    unavailableTimeSlot: any;
    slotDetail: string;
    countryId: any;
    stateId: any;
    cityId: any;
    countryName: string;
    stateName: string;
    cityName: string;
    status:string;
    lat:any;
    long:any;
    zipCode:any;
    admission:string;
    eoc:string;
  }